+++ mon-entreprise/source/components/utils/useInitialRender.tsx
@@ -1,0 +1,28 @@
+import {
+	createContext,
+	ReactNode,
+	useContext,
+	useEffect,
+	useMemo,
+	useState,
+} from 'react'
+
+const InitialRenderContext = createContext(true)
+
+export function WatchInitialRender(props: { children: ReactNode }) {
+	const [initialRender, setInitialRender] = useState(true)
+	useEffect(() => {
+		setInitialRender(false)
+	}, [])
+	return (
+		<InitialRenderContext.Provider value={initialRender}>
+			{props.children}
+		</InitialRenderContext.Provider>
+	)
+}
+
+export function useInitialRender() {
+	const initialRender = useContext(InitialRenderContext)
+	const unChangedInitialRender = useMemo(() => initialRender, [])
+	return unChangedInitialRender
+}
